+++ packages/cozy-app-publish/cozy-app-publish.js
@@ -5,24 +5,13 @@
 const commander = require('commander')
 const colorize = require('./utils/colorize')
 const scripts = require('./index')
+const capitalize = require('lodash/capitalize')
 
 const pkg = require('./package.json')
 
 const MODES = {
   TRAVIS: 'travis',
   MANUAL: 'manual'
-}
-
-var currentNodeVersion = process.versions.node
-var semver = currentNodeVersion.split('.')
-var major = semver[0]
-
-if (major < 6) {
-  console.error(
-    colorize.red(`You are running Node v${currentNodeVersion}.
-    cozy-app-publish requires Node v6 minimum, please update you version of Node.`)
-  )
-  process.exit(1)
 }
 
 const program = new commander.Command(pkg.name)
@@ -38,7 +27,7 @@ const program = new commander.Command(pkg.name)
   )
   .option(
     '--build-dir <relative-path>',
-    'Path fo the build directory relative to the current directory (default ./build)'
+    'Path of the build directory relative to the current directory (default ./build)'
   )
   .option('--build-url <url>', 'URL of the application archive')
   .option(
@@ -56,6 +45,10 @@ const program = new commander.Command(pkg.name)
   .option(
     '--postpublish <script-path>',
     'Hook to process parameters just after publishing, typically to deploy app'
+  )
+  .option(
+    '--tag-prefix <tag-prefix>',
+    'When publishing from a monorepo, only consider tags with tagPrefix, ex: cozy-banks/1.0.1.'
   )
   .option('--yes', 'Force confirmation when publishing manually')
   .option(
@@ -79,19 +72,7 @@ const handleError = error => {
 }
 
 try {
-  publishApp({
-    token: program.token,
-    buildDir: program.buildDir,
-    buildUrl: program.buildUrl,
-    buildCommit: program.buildCommit,
-    manualVersion: program.manualVersion,
-    prepublishHook: program.prepublish,
-    postpublishHook: program.postpublish,
-    yes: program.yes,
-    registryUrl: program.registryUrl,
-    space: program.space,
-    verbose: program.verbose
-  }).catch(handleError)
+  publishApp(program).catch(handleError)
 } catch (error) {
   handleError(error)
 }
@@ -107,45 +88,26 @@ function _getPublishMode() {
 
 async function publishApp(cliOptions) {
   const publishMode = _getPublishMode()
-  if (publishMode === MODES.TRAVIS) {
-    console.log()
-    console.log(`${colorize.bold('Travis')} ${colorize.blue('publish mode')}`)
-    console.log()
-    try {
-      await scripts.travis({
-        registryToken: cliOptions.token,
-        buildDir: cliOptions.buildDir,
-        buildCommit: cliOptions.buildCommit,
-        buildUrl: cliOptions.buildUrl,
-        prepublishHook: cliOptions.prepublishHook,
-        postpublishHook: cliOptions.postpublishHook,
-        registryUrl: cliOptions.registryUrl,
-        spaceName: cliOptions.space,
-        verbose: cliOptions.verbose
-      })
-    } catch (error) {
-      console.error(`↳ ❌  ${error.message}`)
-      process.exit(1)
-    }
-  } else if (publishMode === MODES.MANUAL) {
-    console.log()
-    console.log(`${colorize.bold('Manual')} ${colorize.blue('publish mode')}`)
-    console.log()
-    scripts.manual({
-      registryToken: cliOptions.token,
-      buildDir: cliOptions.buildDir,
-      appBuildUrl: cliOptions.buildUrl,
-      manualVersion: cliOptions.manualVersion,
-      prepublishHook: cliOptions.prepublishHook,
-      postpublishHook: cliOptions.postpublishHook,
-      yes: cliOptions.yes,
-      registryUrl: cliOptions.registryUrl,
-      spaceName: cliOptions.space,
-      verbose: cliOptions.verbose
-    })
-  } else {
-    // no modes found or detected (should not happen since we have default mode)
-    console.log()
-    throw new Error('No modes found or detected. Publishing failed.')
-  }
+  const publishFn = scripts[publishMode]
+
+  console.log()
+  console.log(
+    `${colorize.bold(capitalize(publishMode))} ${colorize.blue('publish mode')}`
+  )
+  console.log()
+  return publishFn({
+    appBuildUrl: cliOptions.buildUrl,
+    buildCommit: cliOptions.buildCommit,
+    buildDir: cliOptions.buildDir,
+    buildUrl: cliOptions.buildUrl,
+    manualVersion: cliOptions.manualVersion,
+    postpublishHook: cliOptions.postpublish,
+    prepublishHook: cliOptions.prepublish,
+    registryToken: cliOptions.token,
+    registryUrl: cliOptions.registryUrl,
+    spaceName: cliOptions.space,
+    tagPrefix: cliOptions.tagPrefix,
+    verbose: cliOptions.verbose,
+    yes: cliOptions.yes
+  })
 }