+++ packages/cozy-harvest-lib/src/components/TriggerManager.jsx
@@ -56,7 +56,11 @@ export class TriggerManager extends Component {
       t
     } = this.props
 
-    const { account } = this.state
+    const { account, trigger } = this.state
+
+    if (trigger) {
+      return trigger
+    }
 
     let folder
 
@@ -72,7 +76,7 @@ export class TriggerManager extends Component {
       await addReferencesTo(konnector, [folder])
     }
 
-    const trigger = await createTrigger(
+    return await createTrigger(
       triggers.buildAttributes({
         account,
         cron: cron.fromKonnector(konnector),
@@ -80,12 +84,6 @@ export class TriggerManager extends Component {
         konnector
       })
     )
-
-    this.setState({
-      trigger
-    })
-
-    return trigger
   }
 
   /**
@@ -96,6 +94,7 @@ export class TriggerManager extends Component {
   async handleAccountSaveSuccess(account) {
     this.setState({ account })
     const trigger = await this.ensureTrigger()
+    this.setState({ trigger })
     return await this.launch(trigger)
   }
 
@@ -159,10 +158,10 @@ export class TriggerManager extends Component {
     })
 
     if (['queued', 'running'].includes(job.state)) {
-      return onLoginSuccess(trigger)
+      return typeof onLoginSuccess === 'function' && onLoginSuccess(trigger)
     }
 
-    return onSuccess(trigger)
+    return typeof onSuccess === 'function' && onSuccess(trigger)
   }
 
   render() {
@@ -199,7 +198,7 @@ TriggerManager.propTypes = {
   statDirectoryByPath: PropTypes.func,
   waitForLoginSuccess: PropTypes.func.isRequired,
   // hooks
-  onLoginSuccess: PropTypes.func.isRequired,
+  onLoginSuccess: PropTypes.func,
   onSuccess: PropTypes.func
 }